refactor: Remove maxCount from X-axis label generation
This commit removes the `maxCount` parameter from the `generateAxisXLabels` function in `DateUtils` and its usages in `WeatherGraphBarsCard` and `WeatherGraphLineCard`.

The `maxCount` and `labelWidthFactor` were used to limit the number of labels displayed on the X-axis based on the available width. This logic has been removed, allowing the system to determine the optimal number of labels.